fix: Remote edit kicks out of Mermaid editing state
closes #11240
